chore: rename dir of exploit mitigations plugin to fit plugin name
