+++ src/plugins/analysis/exploit_mitigations/code/exploit_mitigations.py
