+++ src/plugins/analysis/exploit_mitigations/test/data/Hallo_SafeStack
