+++ src/plugins/analysis/exploit_mitigations/test/test_plugin_exploit_mitigations.py
@@ -4,7 +4,7 @@ import pytest
 
 from objects.file import FileObject
 
-from ..code.checksec import (
+from ..code.exploit_mitigations import (
     AnalysisPlugin,
     check_canary,
     check_clang_cfi,
